pyproject.toml
[project]
name = "unikraft-cloud-platform"
version = "v1.0.0"
description = "A client library for accessing Unikraft Cloud Platform API"
authors = []
requires-python = ">=3.9,<4.0"
readme = "README.md"
dependencies = [
    "httpx>=0.23.0,<0.29.0",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0,<3",
]

[tool.uv.build-backend]
module-name = "unikraft_cloud_platform"
module-root = ""

[build-system]
requires = ["uv_build>=0.8.0,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
